backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
finnhub-python==2.4.14
numpy==1.23.5
gunicorn==20.1.0